cloud-ai-starter-projects/template-3-container-serverless-journal/services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
boto3==1.35.93
python-jose[cryptography]==3.3.0
httpx==0.27.2
pydantic==2.10.3
